Updated deprecated files
